Driving Code
Made some updates to how driving the robot works. The control mode switches to percent output when going down, and resets the position of the evelator to zero when pressed. The start and select button reverted back to the original plan of having those two buttons controlling the elevator by percent output. Hatch exit will not work as DriveSystem is always required by driveVelocity. The locations of where the elevator goes for each hatch height has decreaased by 15 inches as we did not account for the hatch being higher than the intake. Things that need to be worked on include making switching from hatch to ball work, and trying to get a string potentiometer that Mr Cutter is going to add to the robot to tell us where on the elevator we are currently at.